LayoutTests/platform/chromium/plugins/inner-html-display-none-expected.txt

ALERT: function testCallback() { [native code] }
